fix: restore typed module in outline declarations
